fix: quote wallpaper path arg in switch-wallpaper.sh

+++ scripts/dot_config/scripts/switch-wallpaper.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 wallpaper="$HOME/.local/share/backgrounds/background"
-if [ -e $1 ]; then
+if [ -n "$1" ] && [ -e "$1" ]; then
     cp "$1" $wallpaper
     $HOME/.config/scripts/flash-swww-wallpaper.sh
 fi
